AmqpPublishEvent

+++ src/main/scala/io/gatling/amqp/data/PublishAcked.scala
@@ -1,3 +1,3 @@
 package io.gatling.amqp.data
 
-case class PublishAcked(no: Long, multiple: Boolean)
+case class PublishAcked(no: Int, multiple: Boolean)
